XMLHttpRequest/open-url-multi-window-2.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: open() resolving URLs (multi-Window; 2; evil)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    function init(){ // called from page inside IFRAME
        test(function() {
          var client = new self[0].XMLHttpRequest()
          document.body.removeChild(document.getElementsByTagName("iframe")[0])
          assert_throws("INVALID_STATE_ERR", function() { 
            client.open("GET", "folder.txt") 
          }, "open() when associated document's IFRAME is removed")
          });
      }
    </script>
    <iframe src="resources/init.htm"></iframe>
 </body>
</html>
